+++ themes-default/slim/src/store/modules/config/system.js
@@ -62,7 +62,8 @@ const state = {
         lastRead: null,
         latest: null,
         unread: null
-    }
+    },
+    ffprobeVersion: null
 };
 
 const mutations = {